Generalize MeshData into RenderMesh
We're pulling MeshData out of the render_gl infrastructure. It will serve
as the interface for extracting mesh data (OBJ for now) for multiple
render engine implementations. To do this:

1. Extract the struct and functions from render_gl/internal_shape_meshes.*
   into its own render/render_mesh.* shared locale.
2. Names have been changed. MeshData --> RenderMesh (it will eventually
   also carry material information). The LoadMeshFromObj has been
   elaborated to differentiate other OBJ parsing targets.
3. The RenderMesh is not longer OpenGl specific; so we've removed the
   strong numerical types for GLfloat and the like.
   - This requires the consumers of the erstwhile MeshData to change how it
     handles the types.
