+++ geometry/render/BUILD.bazel
@@ -17,6 +17,7 @@ drake_cc_package_library(
         ":render_engine",
         ":render_label",
         ":render_material",
+        ":render_mesh",
     ],
 )
 
@@ -78,6 +79,18 @@ drake_cc_library(
     ],
 )
 
+drake_cc_library(
+    name = "render_mesh",
+    srcs = ["render_mesh.cc"],
+    hdrs = ["render_mesh.h"],
+    interface_deps = [
+        "//common:essential",
+    ],
+    deps = [
+        "@tinyobjloader",
+    ],
+)
+
 # === test/ ===
 
 filegroup(
@@ -127,6 +140,17 @@ drake_cc_googletest(
         ":render_material",
         "//common/test_utilities:diagnostic_policy_test_base",
         "//geometry:geometry_roles",
+    ],
+)
+
+drake_cc_googletest(
+    name = "render_mesh_test",
+    data = [":test_models"],
+    deps = [
+        ":render_mesh",
+        "//common:find_resource",
+        "//common/test_utilities:eigen_matrix_compare",
+        "//common/test_utilities:expect_throws_message",
     ],
 )
