+++ geometry/render_gl/BUILD.bazel
@@ -79,6 +79,7 @@ drake_cc_library_ubuntu_only(
         ":internal_texture_library",
         ":render_engine_gl_params",
         "//geometry/render:render_engine",
+        "//geometry/render:render_mesh",
         "//systems/sensors:image",
     ],
 )
@@ -164,6 +165,7 @@ drake_cc_library_ubuntu_only(
     hdrs = ["internal_shape_meshes.h"],
     interface_deps = [
         ":internal_opengl_context",
+        "//geometry/render:render_mesh",
     ],
     deps = [
         "//common:essential",
