+++ geometry/render_gl/internal_shape_meshes.h
@@ -5,60 +5,12 @@
 
 #include <Eigen/Dense>
 
-#include "drake/geometry/render_gl/internal_opengl_includes.h"
+#include "drake/geometry/render/render_mesh.h"
 
 namespace drake {
 namespace geometry {
 namespace render_gl {
 namespace internal {
-
-/* The data representing a mesh. The triangle mesh is defined by `indices`. Row
- t represents a triangle by the triplet of vertex indices: tᵥ₀, tᵥ₁, tᵥ₂. The
- indices map into the rows of `positions`, `normals`, and `uvs`. I.e., for
- vertex index v, the position of that vertex is at `positions.row(v)`, its
- corresponding normal is at `normals.row(v)`, and its texture coordinates are at
- `uvs.row(v)`.
-
- For now, all vertex quantities (`positions`, `normals` and `uvs`) are
- guaranteed (as well as the `indices` data). In the future, `uvs` may become
- optional.  */
-struct MeshData {
-  Eigen::Matrix<GLfloat, Eigen::Dynamic, 3, Eigen::RowMajor> positions;
-  Eigen::Matrix<GLfloat, Eigen::Dynamic, 3, Eigen::RowMajor> normals;
-  Eigen::Matrix<GLfloat, Eigen::Dynamic, 2, Eigen::RowMajor> uvs;
-  Eigen::Matrix<GLuint, Eigen::Dynamic, 3, Eigen::RowMajor> indices;
-
-  /** See docs for `has_tex_coord` below.  */
-  static constexpr bool kHasTexCoordDefault{true};
-
-  /** This flag indicates that this mesh has texture coordinates to support
-   maps.
-   If True, the values of `uvs` will be nontrivial.
-   If False, the values of `uvs` will be all zeros, but will still have the
-   correct size.  */
-  bool has_tex_coord{kHasTexCoordDefault};
-};
-
-/* Loads a mesh's vertices and indices (faces) from an OBJ description given
- in the input stream. It does not load textures. Note that while this
- functionality seems similar to ReadObjToTriangleSurfaceMesh, RenderEngineGl
- cannot use TriangleSurfaceMesh. Rendering requires normals and texture
- coordinates; TriangleSurfaceMesh was not designed with those quantities in
- mind.
-
- If no texture coordinates are specified by the file, it will be indicated in
- the returned MeshData. See MeshData::has_tex_coord for more detail.
-
- @throws std::exception if a) tinyobj::LoadObj() fails, (b) there are no faces
-                           or normals, c) faces fail to reference normals, or d)
-                           faces fail to reference the texture coordinates if
-                           they are present.  */
-MeshData LoadMeshFromObj(std::istream* input_stream,
-                         const std::string& filename = "from_string");
-
-/* Overload of LoadMeshFromObj that reads the OBJ description from the given
- file. */
-MeshData LoadMeshFromObj(const std::string& filename);
 
 // TODO(SeanCurtis-TRI): Provide a geodesic sphere (or a tessellation like that
 //  produced for hydroelastics).
@@ -78,8 +30,8 @@ MeshData LoadMeshFromObj(const std::string& filename);
                             with the sphere equator.
  @pre `longitude_bands` >= 3 and `latitude_bands` >= 2 (otherwise the sphere
       will have no volume).  */
-MeshData MakeLongLatUnitSphere(int longitude_bands = 50,
-                               int latitude_bands = 50);
+geometry::internal::RenderMesh MakeLongLatUnitSphere(int longitude_bands = 50,
+                                                     int latitude_bands = 50);
 
 /* Creates an OpenGL-compatible mesh representation of a unit cylinder; its
  radius and height are equal to 1. It is centered on the origin of its canonical
@@ -105,7 +57,8 @@ MeshData MakeLongLatUnitSphere(int longitude_bands = 50,
                     cylinder.
  @pre `num_strips` >= 3 (otherwise the cylinder will have no volume).
  @pre `num_bands` >= 1.  */
-MeshData MakeUnitCylinder(int num_strips = 50, int num_bands = 1);
+geometry::internal::RenderMesh MakeUnitCylinder(int num_strips = 50,
+                                                int num_bands = 1);
 
 /* Creates an OpenGL-compatible mesh reprepsenting a square patch. The patch
  has edge length `measure` units long. The square patch is defined lying on the
@@ -121,7 +74,8 @@ MeshData MakeUnitCylinder(int num_strips = 50, int num_bands = 1);
  two triangles).
  @pre `measure` > 0
  @pre `resolution >= 1`. */
-MeshData MakeSquarePatch(GLfloat measure = 200, int resolution = 1);
+geometry::internal::RenderMesh MakeSquarePatch(double measure = 200,
+                                               int resolution = 1);
 
 /* Creates an OpenGL-compatible mesh representation of the unit box - all edges
  are length 1. The box is centered on the origin of its canonical frame C with
@@ -139,7 +93,7 @@ MeshData MakeSquarePatch(GLfloat measure = 200, int resolution = 1);
 
 <!-- TODO(SeanCurtis-TRI): consider offering subdivisions if per-vertex
     properties yield undesirable artifacts for large boxes.  --> */
-MeshData MakeUnitBox();
+geometry::internal::RenderMesh MakeUnitBox();
 
 /* Creates an OpenGL-compatible mesh representation of a capsule with the given
  `radius` and `length`. The capsule is centered on the origin of its canonical
@@ -167,7 +121,8 @@ MeshData MakeUnitBox();
  @param length    The length of the cylindrical barrel.
  @pre `samples` >= 3 (otherwise the capsule will have no volume).
  @pre radius > 0 and length > 0.  */
-MeshData MakeCapsule(int samples, double radius, double length);
+geometry::internal::RenderMesh MakeCapsule(int samples, double radius,
+                                           double length);
 
 }  // namespace internal
 }  // namespace render_gl
